requirements.txt
Django==3.2
django-cors-headers==3.10.0
django-filter==2.4.0
django-ckeditor==6.4.0
django-jazzmin==2.6.0
djangorestframework==3.12.4
pillow==10.4.0
psycopg2==2.9.1
python-dateutil==2.8.2
python-dotenv==0.19.1
random-username==1.0.2
requests==2.26.0